changed file structures to match template site
